Some minor changes made in the book's sources.
To be able to use the actuator endpoints, a separate docker container
is launched. To simplify using this, a explicit network name is added to the
Dockerfiles.
The container is not added to the Dockerfiles but will be started,
when the information is needed (e.g. in the test-em-all.sh script).